+++ server/src/api/analytics/funnels/getFunnel.ts
@@ -2,8 +2,8 @@ import { FilterParams } from "@rybbit/shared";
 import { FastifyReply, FastifyRequest } from "fastify";
 import SqlString from "sqlstring";
 import { clickhouse } from "../../../db/clickhouse/clickhouse.js";
-import { getTimeStatement, patternToRegex, processResults } from "../utils.js";
-import { getFilterStatement } from "../getFilterStatement.js";
+import { getTimeStatement, patternToRegex, processResults } from "../utils/utils.js";
+import { getFilterStatement } from "../utils/getFilterStatement.js";
 
 type FunnelStep = {
   value: string;
