+++ server/src/api/analytics/getErrorBucketed.ts
@@ -2,10 +2,10 @@ import { FilterParams } from "@rybbit/shared";
 import { FastifyReply, FastifyRequest } from "fastify";
 import SqlString from "sqlstring";
 import { clickhouse } from "../../db/clickhouse/clickhouse.js";
-import { validateTimeStatementFillParams } from "./query-validation.js";
-import { getTimeStatement, TimeBucketToFn, bucketIntervalMap, processResults } from "./utils.js";
+import { validateTimeStatementFillParams } from "./utils/query-validation.js";
+import { getTimeStatement, TimeBucketToFn, bucketIntervalMap, processResults } from "./utils/utils.js";
 import { TimeBucket } from "./types.js";
-import { getFilterStatement } from "./getFilterStatement.js";
+import { getFilterStatement } from "./utils/getFilterStatement.js";
 
 function getTimeStatementFill(params: FilterParams, bucket: TimeBucket) {
   const { params: validatedParams, bucket: validatedBucket } = validateTimeStatementFillParams(params, bucket);
