+++ server/src/api/analytics/getJourneys.ts
@@ -1,8 +1,8 @@
 import { FilterParams } from "@rybbit/shared";
 import { FastifyReply, FastifyRequest } from "fastify";
 import { clickhouse } from "../../db/clickhouse/clickhouse.js";
-import { getFilterStatement } from "./getFilterStatement.js";
-import { getTimeStatement } from "./utils.js";
+import { getFilterStatement } from "./utils/getFilterStatement.js";
+import { getTimeStatement } from "./utils/utils.js";
 
 export const getJourneys = async (
   request: FastifyRequest<{
