+++ server/src/api/analytics/getRetention.ts
@@ -1,6 +1,6 @@
 import { FastifyReply, FastifyRequest } from "fastify";
 import { clickhouse } from "../../db/clickhouse/clickhouse.js";
-import { processResults } from "./utils.js";
+import { processResults } from "./utils/utils.js";
 
 // Define the expected shape of a single data row from the query
 interface RetentionDataRow {
