+++ server/src/api/analytics/goals/getGoals.ts
@@ -3,10 +3,10 @@ import { db } from "../../../db/postgres/postgres.js";
 import { goals } from "../../../db/postgres/schema.js";
 import { clickhouse } from "../../../db/clickhouse/clickhouse.js";
 import { eq, desc, asc, sql } from "drizzle-orm";
-import { getTimeStatement, processResults, patternToRegex } from "../utils.js";
+import { getTimeStatement, processResults, patternToRegex } from "../utils/utils.js";
 import SqlString from "sqlstring";
 import { FilterParams } from "@rybbit/shared";
-import { getFilterStatement } from "../getFilterStatement.js";
+import { getFilterStatement } from "../utils/getFilterStatement.js";
 
 // Types for the response
 interface GoalWithConversions {
