+++ server/src/api/analytics/performance/getPerformanceTimeSeries.ts
@@ -1,11 +1,11 @@
 import { FastifyReply, FastifyRequest } from "fastify";
 import { clickhouse } from "../../../db/clickhouse/clickhouse.js";
-import { getTimeStatement, processResults, TimeBucketToFn, bucketIntervalMap } from "../utils.js";
+import { getTimeStatement, processResults, TimeBucketToFn, bucketIntervalMap } from "../utils/utils.js";
 import SqlString from "sqlstring";
-import { validateTimeStatementFillParams } from "../query-validation.js";
+import { validateTimeStatementFillParams } from "../utils/query-validation.js";
 import { TimeBucket, PerformanceTimeSeriesPoint } from "../types.js";
 import { FilterParams } from "@rybbit/shared";
-import { getFilterStatement } from "../getFilterStatement.js";
+import { getFilterStatement } from "../utils/getFilterStatement.js";
 
 function getTimeStatementFill(params: FilterParams, bucket: TimeBucket) {
   const { params: validatedParams, bucket: validatedBucket } = validateTimeStatementFillParams(params, bucket);
