+++ server/src/api/analytics/utils/query-validation.ts
