+++ server/src/api/uptime/getMonitorUptimeBuckets.ts
@@ -5,7 +5,7 @@ import { clickhouse } from "../../db/clickhouse/clickhouse.js";
 import { db } from "../../db/postgres/postgres.js";
 import { member, uptimeMonitors } from "../../db/postgres/schema.js";
 import { getSessionFromReq } from "../../lib/auth-utils.js";
-import { processResults } from "../analytics/utils.js";
+import { processResults } from "../analytics/utils/utils.js";
 
 // Query params schema
 const getMonitorUptimeBucketsQuerySchema = z.object({
