+++ server/src/api/uptime/getMonitors.ts
@@ -5,7 +5,7 @@ import { db } from "../../db/postgres/postgres.js";
 import { uptimeMonitors, uptimeMonitorStatus, member } from "../../db/postgres/schema.js";
 import { clickhouse } from "../../db/clickhouse/clickhouse.js";
 import { getSessionFromReq } from "../../lib/auth-utils.js";
-import { processResults } from "../analytics/utils.js";
+import { processResults } from "../analytics/utils/utils.js";
 import { getMonitorsQuerySchema, type GetMonitorsQuery } from "./schemas.js";
 
 interface GetMonitorsRequest {
